scrapper/requirements.txt
appdirs==1.4.4
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.1
cssselect==1.1.0
dnspython==2.2.1
fake-useragent==0.1.11
html5lib==1.1
idna==3.3
importlib-metadata==4.12.0
lxml==4.9.1
parse==1.19.0
Pillow==9.2.0
psutil==5.9.2
pyee==8.2.2
pymongo==4.2.0
pyppeteer==1.0.2
pyquery==1.4.3
requests==2.28.1
requests-html==0.10.0
six==1.16.0
soupsieve==2.3.2.post1
tqdm==4.64.1
urllib3==1.26.12
w3lib==2.0.1
webencodings==0.5.1
websockets==10.3
zipp==3.8.1
